Removed binary files from the repository so template.iso was removed. README.md was edited to capture additional steps to make up for that and the fuzzer-master was edited to fix cases with iso fuzzing without a template file.

+++ core/fuzzer-master.ps1
@@ -110,15 +110,25 @@ if ($VirtualDiskType) {
     }
     $fsGeneratorPath = Join-Path $ScriptPath "fs-generator.ps1"
     if ($VirtualDiskType.GetType().Name -eq "String") {
-        Copy-Item "..\data\template.$VirtualDiskType" $fsCorpus
         $fsTemplate = Join-Path $fsCorpus "template.$VirtualDiskType"
+        if (Test-Path -Path "..\data\template.$VirtualDiskType") {
+            Copy-Item "..\data\template.$VirtualDiskType" $fsCorpus
+        }
+        else {
+            New-Item -Path $fsTemplate -Value "Test" -ItemType File -Force > $null
+        }
         $fscommand = "-HostOutputDir ""$HostOutputDir"" -VirtualDiskType $VirtualDiskType"
         $fsGenPid = Start-HelperProcess -script $fsGeneratorPath -commandline $fscommand
     } else {
         $fsTemplate = New-Object System.Collections.ArrayList
         $fscommands = New-Object System.Collections.ArrayList
         foreach ($format in $VirtualDiskType) {
-            Copy-Item "..\data\template.$format" $fsCorpus
+            if (Test-Path -Path "..\data\template.$VirtualDiskType") {
+                Copy-Item "..\data\template.$format" $fsCorpus
+            }
+            else {
+                New-Item -Path (Join-Path $fsCorpus "template.$format") -Value "Test" -ItemType File -Force > $null
+            }
             $fsTemplate.Add((Join-Path $fsCorpus "template.$format")) > $null
             $fscommand = "-HostOutputDir ""$HostOutputDir"" -VirtualDiskType $format"
             $fscommands.Add($fscommand) > $null
